refactor: move zoom to extent logic to separate hook
